+++ src/test/java/school/redrover/model/FolderProjectPage.java
@@ -19,12 +19,6 @@ public class FolderProjectPage extends BaseProjectPage {
     @FindBy(tagName = "h1")
     private WebElement pageHeading;
 
-    @FindBy(xpath = "//span[contains(text(),'Rename')]/../.")
-    private WebElement renameButtonLeft;
-
-    @FindBy(css = "h1")
-    private WebElement pageTopic;
-
     @FindBy(css = ".empty-state-section")
     private WebElement emptyStateSection;
 
@@ -71,12 +65,6 @@ public class FolderProjectPage extends BaseProjectPage {
 
     public FolderRenamePage clickOnRenameButton() {
         renameButton.click();
-
-        return new FolderRenamePage(getDriver());
-    }
-
-    public FolderRenamePage clickOnRenameButtonLeft() {
-        renameButtonLeft.click();
 
         return new FolderRenamePage(getDriver());
     }
