+++ src/test/java/school/redrover/model/HomePage.java
@@ -41,9 +41,6 @@ public class HomePage extends BasePage {
     @FindBy(css = "[href*='rename']")
     private WebElement renameFromDropdown;
 
-    @FindBy(css = "div.jenkins-dropdown")
-    private WebElement dropdownMenu;
-
     @FindBy(xpath = "//a[@class='sortheader' and text()='Name']")
     private WebElement columnNameTitle;
 
@@ -250,12 +247,6 @@ public class HomePage extends BasePage {
         return new ViewPage(getDriver());
     }
 
-    public FolderProjectPage clickOnCreatedFolder(String name) {
-        getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//tr[@id='job_" + name + "']/td/a"))).click();
-
-        return new FolderProjectPage(getDriver());
-    }
-
     public HomePage openItemDropdownWithSelenium(String projectName) {
         new Actions(getDriver())
                 .moveToElement(getDriver().findElement(
@@ -354,12 +345,6 @@ public class HomePage extends BasePage {
         return this;
     }
 
-    public CreateNewItemPage clickNewJobFromDashboardBreadcrumbsMenu() {
-        getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("div.tippy-box a[href $= 'newJob']"))).click();
-
-        return new CreateNewItemPage(getDriver());
-    }
-
     public ManageJenkinsPage clickManageFromDashboardBreadcrumbsMenu() {
         manageFromDashboardBreadcrumbsMenu.click();
 
@@ -396,13 +381,6 @@ public class HomePage extends BasePage {
         return new FullStageViewPage(getDriver());
     }
 
-    public List<String> allExistingJobsNames() {
-        return allExistingJobs
-                .stream()
-                .map(WebElement::getText)
-                .toList();
-    }
-
     public List<String> getJobsBeginningFromThisFirstLetters(String firstLetters) {
         return allExistingJobs
                 .stream()
@@ -436,7 +414,6 @@ public class HomePage extends BasePage {
     public boolean isDisplayedWebsiteDropdownItem() {
         return websiteDropdownItem.isDisplayed();
     }
-
 
     public HomePage moveMouseToPassiveViewName() {
         new Actions(getDriver())
@@ -518,7 +495,6 @@ public class HomePage extends BasePage {
         return new MovePage(getDriver());
     }
 
-
     public AboutPage clickAbout() {
         about.click();
 
@@ -574,6 +550,5 @@ public class HomePage extends BasePage {
         dropdownPipelineSyntax.click();
 
         return new PipelineSyntaxPage(getDriver());
-
     }
 }
